Tunneled id-space-merge-3 through master.  Fixed link and run time errors.  Commented out a number of obsolete test drivers from the build and marked them as $$OBSOLETE in the .t.cc files.
Compiles, links and runs on linux with the exception of the some [fiber]_x.t test drivers failing.  These failures appear to happen in when the fiber goes out of scope and the postcondition of vd::put_component() fails.  Can not debug because the test driver(s) that fail when ran from a command line, run to completion in gdb.  The test driver that fails changes each time 'make all' is ran.  Sometimes no test drivers fail.